Removed trailing f on real number outputs

+++ src/interface/stringable.cpp
@@ -195,7 +195,7 @@ string PrintExp::toString() {
     return s;
 }
 string RealExp::toString() {
-    return to_string(val) + "f";
+    return to_string((double) val) + "f";
 }
 string SequenceExp::toString() {
     string s = "";
@@ -300,7 +300,7 @@ string ListVal::toString() {
     s += "]";
     return s;
 }
-string RealVal::toString() { return to_string(val); }
+string RealVal::toString() { return to_string((double) val); }
 string StringVal::toString() { return val; }
 string Thunk::toString() { return val ? val->toString() : ("" + exp->toString() + " | " + env->toString()); }
 string TupleVal::toString() { return "(" + left->toString() + ", " + right->toString() + ")"; }
